+++ examples/react-web-component-demo/src/App.jsx
@@ -1,0 +1,13 @@
+import './App.css'
+import PivotHeadDemo from './PivotHeadDemo'
+
+function App() {
+
+  return (
+    <>
+     <PivotHeadDemo />
+    </>
+  )
+}
+
+export default App