+++ examples/simple-js-demo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # simple-js-demo
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mindfiredigital/pivothead@1.1.0
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mindfiredigital/pivothead@1.0.8
+
 ## 1.0.6
 
 ### Patch Changes